Implemented bounds
Updated constructors & addTree() Methods to properly handle implementation of start_ and end_ values.

These value's represent the start and end value's of the row. If attempt to add a tree outside of the bounds of the row will automatically adjust the bounds of the row to include this new tree. 

Also implemented treeLocations_ list more fully: list now should properly handle containing the position of a tree within a row.