Create Leaf
Moved "Content" String and "VisibleDate" int out of Tree class into new Leaf class. 

Each leaf contains a single String (though this string can be as long as a user wants). and a single VisibleDate. The VisibleDate corresponds to the "date" at which the leaf becomes visible on the Tree.