Processing Rows
Started implementation of multiple rows, and multiple trees with the same plant date.

To handle this implemented processRow method, which takes 
a row, and loads it into an array of arrays. The first entry of each sub array is the "plant date" with all following entries being trees with the appropriate plant date. 

In future this structure may become more deeply ingrained into the shape of a Row object, but for now it exists as a way to take the loosely organized row structure, and convert it into one more suited for rendering purposes.