Semi working pre Rowrework
have come to the decision that the current implementation of the row data structure lacks alot of useful structure.

The current structure is too closely based on the general row structure of the JSON object.

Planed update is more reminescent of the array structure present in the "ProcessRow" method. At the moment I'm doing far to much individual processing in methods, instead of storing relevant info inside data structures that are then passed around to be worked on.